Added additional options for choosing modelnames
Added "Model Options" section to web ui (next to fit options)
which allows specifying individual modelnames that do *not*
get parsed as a modeltree in the usual manner. They are also
separate from the analyses saved in the database so that you can,
for example, enter a specific model (or list of comma-separated models)
and then compare them to several different analyses without needing
to add them to the modeltree of each other analysis.

Additionally, extra analysis names can be specified in order to
add all modelnames from that analysis onto the modelnames for
the current analysis (skipping duplicates). For example,
if 'Analysis 2' is specified as an extra analysis while looking
at 'Analysis 1', then all of the modelnames for both Analysis 1
and Analysis 2 will be visible, without making any changes to the
modeltree for either analysis.